Manual pgBackRest Backups
It is now possible for end-users to manually trigger pgBackRest backup
Jobs for a specific pgBackRest repository as defined within the their
PostgresCluster spec. Using the "spec.archive.pgbackrest.manual" field
within the spec, the end-user can specify a repo that they would like
to backup, along with various pgBackRest command-line options to
include when running the "pgbackrest backup" command.

To trigger a manual backup (specifically once the pertinent details
have been populated within the "manual" section in the spec, as
described above), the user will annotate the PostgresCluster using the
"postgres-operator.crunchydata.com/pgbackrest-backup" annotation.  The
value provided for the annotation should be a unique identifier, which
is used by the Operator to properly track the status of the Job.

The status of a manual backup Job is tracked in the status of the
PostgresCluster, specifically providing insight into the start and
successful completion of the Job, along with any Pods that failed or
succeeded when running the Job.  Additionally, a
"PGBackRestManualBackupSuccessful" condition is populated to indicate
whether or not the last manual backup Job completed successfully.

[ch10760]