Allow a separate WAL volume to be added
PostgreSQL commits transactions by storing changes in its write-ahead
log. The contents of the log are applied to data files (containing
tables and indexes) later as part of a checkpoint.

The way WAL files are accessed and utilized often differs from that of
data files. In high-performance situations, it can desirable to put WAL
files on separate storage.

At this time, adding a WAL volume is a one-way operation. It can be done
when the PostgresCluster is created or anytime thereafter, but it cannot
yet be removed. That ability will come in a later commit.

Issue: [ch10778]
See: https://www.postgresql.org/docs/current/wal.html