Custom Pod & Node Affinity
Allows users to define corev1.Affinity rules for various Pods comprising
a PostgreSQL cluster. This includes individual instances and the
dedicated repo host (if enabled).

The corev1.Affinity object defined in the PostgresInstanceSetSpec or
DedicatedRepo fields of the PostgresClusterSpec will be automatically
added to the instance and/or dedicated repo host pod templates.