Reuse PVC if StatefulSet is Deleted
Existing PVC's are now re-used in the event that the operator detects
an existing PVC in the environment for a specific instance set when
scaling-up that specific instance set.  For instance, if the
StatefulSet associated with an existing instance is accidentally
deleted, it will now be recreated with the same name and using the same
PVC, instead of the operator reconciling a new StatefulSet (with a new
name) and a new PVC.  This therefore ensures the effective reuse of
existing volumes when healing a cluster.