Only Reuse PGDATA PVC if WAL PVC is Also Found
If an instance set has a seperate WAL volume enabled by configuring
'walVolumeClaimSpec' within the PostgresCluster spec, then an existing
PGDATA PVC will only be reused (e.g. in the event an instance
StatefulSet is inadvertently deleted) if a corresponding WAL PVC for
that same instance can be found.  Otherwise, if both cannot be found,
a new instance StatefulSet and PVC will be reconciled.