Consistent Use of Observed Instances when Scaling
The scale-up and scale-down functions utilized to reconcile the
instances for a PostgreSQL cluster now utilize the "observed
instances" that are captured prior to PostgresCluster reconciliation.
This is done to ensure a consistent view of cluster resources when
reconciling the various instance sets defined in the spec, and
specifically prevents the scale-up/scale-down functions from fighting
with one another when determining whether or not instance set resources
(e.g. StatefulSet's or PVC's) need to be created or deleted.