Ensure "vendor" label is present on PGO objects
Though PGO leverages a lot of stable naming, this label is a
helpful identifier among the objects that PGO is managing. Some
derived objects (e.g. Secrets from SAs) do not get the label, but
they are tightly coulpled to the object that does get the label.

Most objects had the "vendor" label already, but this should bring
everything inline.

Issue: #2470