Skip some tests with check-envtest-existing
Tests for reconciling resources that require a postgrescluster object as
an owner are failing due to the garbage collector. The collector will
delete these resource before the test completes when the owner does not
exist. Until we find a better solution for running these tests without
creating the owning object, we are skipping them when running against a
live cluster.