Allow Restoring From Another PostgresCluster
When creating a new PostgresCluster, it is now possible to specify
another existing PostgresCluster within the same namespace as the
data source for the new cluster.  This will specifically result in
the creation of a pgBacKRest restore Job that will pre-populate the
contents of the PostgreSQL data directory for the new cluster.  This
enables users to not only leverage the pgBackRest backups for other
existing clusters to perform full restores when creating new clusters,
but other types of restores (e.g. point-in-time restores) can be
performed as well.

In order to specify a PostgresCluster cluster data source, users
will populate the "spec.dataSource.postgresCluster" section of the
PostgresCluster custom resource, specifically with the name of
the PostgresCluster to use as the data source, as well as the name of
pgBackRest repository within that cluster to use.   Additionally,
pgBackRest command line options can also be specified to further
configure the restore Job as needed.  If either the cluster, repo
or options are invalid, an "InvalidDataSource" event will be
generated and recorded for the PostgresCluster.

Data source reconciliation is completed prior to the reconciliation
of any other instance, pgBackRest and/or PgBouncer resources.
Additionally, data source reconciliation is designed to run only
once at cluster creation.  Therefore, a condition is set once data
source initialization is complete to prevent the data source
initialization Job from running again.  And even if this condition
were to be removed, if the operator detects that the cluster is
already bootstrapped, the it is assumed that the data source for the
cluster has already be initialized.

[ch11533]