Allows Users to Shutdown and Start Up PostgreSQL Clusters
This update allows a user to shutdown and subsequently start back
up their postgrescluster. This is done by setting the shutdown
option on the postgrescluster spec to true. Once set, an ordered
shutdown occurs, starting with any pgBouncer pods, instance replicas
and repo host pods, then lastly moving to the primary instance pod.
All PVCs, Services, StatefulSets, etc. are left in place. CronJobs
are kept, but suspended.

When the shutdown value is set to false or removed, the process is
reversed so that the first the primary instance pod, the repo host pod
and pgBouncer pod(s) are recreated, followed by the replica pod(s).

This includes internal documentation that describes how the
shutdown and startup code works in detail.