EnableD/Disable exporter sidecar
Allows users to add/remove the exporter sidecar container to PostgreSQL
Pods comprising a PostgreSQL cluster. The PostgresCluster spec has been
updated to include two new main sections, MonitoringSpec and
MonitoringStatus. MonitoringSpec is used to enable/disable the exporter
sidecar and has this format:

  monitoring: pgMonitor: exporter: {}

The Exporter spec expects an image name and supports
resourceRequirements for the container.

The MonitoringStatus is used to keep track of the pgMonitor
configuration installed in the database. The ExporterConfiguration field
is a hash of the sql that has been run in the database. If this field
doesn't match the expected configuration, the configuration will be
updated.

The exporter sidecar container is started using the existing start.sh
script. All environment variables that would be defaulted by the script
are defined in the container spec. The downstream api is mounted to the
database container allowing the exporter to query node metrics.

Once the sidecar is enabled on all instances, reconcile steps are
performed to enable the ccp_monitoring user in the database. The
setup.sql script for Postgrescluster.Spec.Postgresversion is run in the
database and the password for ccp_monitoring user is updated. The
password is stored in a <cluster-name>-monitoring secret that is mounted
to the exporter container as environment variables. Extensions required
by the exporter are also enabled and created in the database.

When the exporter is disabled uninstall steps are performed. Login
permissions for the monitoring user are removed and the monitoring
secret is deleted.

Further work needs to be done to further evaluate the image and scripts
that are used to run the postgres_exporter binary. Running the exporter
binary has not been fully converted from bash to go.