Wire-up Resources & General Restore Updates
Custom resources can now be configured for the pgBackRest restore Job
that is created to initialize a PostgreSQL data directory (e.g. when
using an existing PostgresCluster as the data source for a new
PostgresCluster).

Additionally, condition "DataSourceInitialized" has now been renamed to
"PostgresDataInitialized" to better indicate that it is the PGDATA
directory and contents that are being initialized, while the associated
messages for this condition have also been updated.

Also, during a restore the "target-action" is now also automatically
set to "promote", and a validation warning is recorded if the user
attempts to set this setting via the spec.  This is to ensure the
restore is able to complete according to design, and the resulting
PGDATA directory can then be utilized to bootstrap a new cluster.

And finally, integrates the latest "graceful startup" changes from
#101 to ensure the proper bootstrapping of the cluster using the
PostgreSQL data directory that is pre-populated using a pgBackRest
restore, while also adding/updating various comments to better clarify
various aspects of the restore process.