Configure CRON Jobs to Run pgBackRest Backups
The backup schedules currently defined in the PostgresCluster spec
currently only run a "dummy" Job. This commit updates the behavior so
that the defined pgBackRest backups for the PostgreSQL cluster are
created instead. Full, Differential and Incremental backups are
supported and are defined per repo in the PostgreSQL cluster spec.