Enables Restore In-Place
Users are now able to restore PostgresCluster's in place using the
cluster's physical backups.  Specifically, a "restore" section now
exists within the pgBackRest section of the spec, which allows the
user to define an in-place restore for the cluster (i.e. the
repository that should be utilized for the restore, as well as any
pgBackRest CLI options that should be utilized when running the
"pgbackrest restore" command).  This allows users to perform full
restores, point-in-time restores, etc. to restore their running
PostgreSQL clusters.

To initiate an in-place restore, the user must set the "enabled"
field within the "restore" section of the spec to "true", and then
annotate the PostgresCluster using the "pgbackrest-restore"
annotation.  The value of the annotation should be a unique
identifier (e.g. a timestamp) that can be utilized by the operator
across each restore to track the it's completion, and prevent
additional restores from occurring as the PostgresCluster is
continuously reconciled throughout its lifetime.

This approach not only allows the operator to track the completion
of restores, but also allows the user to explicitly disable in-place
restores for a PostgresCluster when they are not needed (important
since in-place restores are destructive and fully recreate the
database).  This ensures the user can always explicitly ensure an
in-place restore will not run, e.g. to protect against the accidental
loss of PostgresCluster status information (and therefore a loss of
the operator's ability to track whether or not a specific restore has
previously been completed).

During an in-place restore, a "PGBackRestoreProgressing" condition
is created to indicate that an in-place restore is in progress.  This
condition is updated with details about the restore process while
a restore is occurring, and is then removed when the restore is
complete.

It should also be noted that with this commit (which has been
consolidated with the restore process for initializing the PostgreSQL
data directory for a new cluster), users can now use a
PostgresCluster's own existing backups to perform a pgBackRest restore
and initialize the PostgreSQL data directory for that cluster.
Specifically, if a previous pgBackRest repository PV is bound to the
PVC defined for a repository in a new PostgresCluster (e.g. using the
"volumeName" field when defining a repository volume), the user can
define a proper pgBackRest data source in the spec for that specific
repository, and it will the utilized to perform a restore and
initialize the data directory for the cluster.

[ch10761]