Set TargetPort on Primary Service Intent
The 'targetPort' field is now set on the intent that is generated when
reconciling the PostgresCluster's primary service.

This is specifically due to an issue identified in Kubernetes v1.19
and v1.20, in which the SSA patch for the primary service (which
debugging shows includes a default value of '0' for 'targetPort'),
was leading to updates to the Primary Service, even though nothing in
the intent generated by the operator had changed.  This, in turn, was
triggering additional events, and therefore and endless reconciles for
the PostgresCluster.

By explicitly setting 'targetPort', this behavior is no longer seen
in Kubernetes v1.19 and v1.20, and reconciliation for the Primary
Service also continues to work as expected in Kubernetes v1.18 and
v1.21.

[ch11741]