Allows Restoring Across Namespaces
It is now possible to restore (both in-place, or when bootstrapping a
new cluster via a PostgresCluster data source) using a PostgresCluster
in another namespace.  Specifically, users can now define a cluster
namespace when defining the cluster to utilize to either initialize the
data directory for a new PostgrsCluster, or perform an in-place restore
of an existing PostgresCluster.  If no namespace is specified, the
namespace of the PostgresCluster being reconciled is used.

When a restore is requested for a PostgresCluster in another namespace,
the required SSH and pgBackRest configuration is copied from the source
PostgresCluster to the local namespace for the PostgresCluster being
reconciled, where is remains for the duration of the restore (i.e.
following the restore any configuration created for the restore is
removed).

Please note that the Operator must be installed with cluster-level
permissions for cross-namespace restores since the operator must be
able to reach across namespaces to obtain any configuration required to
perform the restore.

[ch11647]