Set statuses for pgBackRest Backup Schedules
The initial pgBackRest backup schedule implementation did not include
relevant statuses for the various states of the scheduled backup
processes. This commit captures the relevant statuses from the Jobs
created by the scheduled pgBackRest backup CronJobs and adds that
information to the PostgreSQL cluster status. Note that the values
will match the current Jobs in the Kubernetes environment, and as Jobs
are cleaned up, those statuses will be removed from the CRD as well.