Standby cluster that replays a pgBackRest archive
A PostgresCluster can now run as a read-only copy of a pgBackRest
repository. WAL files are replayed as they arrive in the repo. To
function correctly, the "spec.standby.enabled" field must be `true` when
the cluster is created, though it can be set or changed to `false` at
any time.

When "spec.standby" is omitted or "spec.standby.enabled" is `false`, the
cluster stops replaying WAL files and starts accepting writes from
client connections. Other changes to the PostgreSQL data kick-off as
well:

- PgBouncer and pgMonitor credentials are reset.
- pgBackRest stanzas are created and scheduled backups resume.

There is currently no way to specify that a repository be read-only.
When the cluster is "promoted" by disabling standby, it will start
pushing WAL files to the same repository it had been fetching from.

Issue: [ch11711]