Only Add Repo Volume If Not Nil
When reconciling pgBackRest repos, the status for a repo is only set
if the corresponding repo volume (PersistentVolumeClaim) reconciled,
and therefore isn't nil.  This prevents a panic from occurring when
reconciliation fails for a repo volume, and the operator then attempts
to set the status for any/all repos configured for the cluster.

[ch11802]