Add go test for image pull secrets on scheduled backup jobs
Because of the overlap between the merge of the PRs for 11701 and 11739,
the go test for checking that the image pull secret is properly set for
the job created by the scheduled backup cronjob has not yet been added.

This commit adds an image pull secret test in line with the existing
tests already in place.