Enable Customization of shared_preload_libraries When Exporter Is Enabled
Currently, when monitoring is turned on, the shared_preload_libraries
get hard coded to 'pg_stat_statements,pgnodemx' because of how the
mandatory parameters are being set by the Operator.

This commit updates the behavior so that user defined shared_preload_libraries
are appended to the existing list, leaving the mandatory libraries in place
while keeping any user defined libraries as well.