New OLM bundle format using Kustomize
OLM prefers packages as "bundle" images these days. The format is
similar to the older "package manifests", but bundles are shipped as
images rather than just a directory. We now generate packages in the
bundle format and validate them and the resulting images using OLM
tools.

Now that we maintain a Kustomize installer, we no longer need to keep a
copy of Deployments, RBAC, etc. specific to OLM. When generating the
package, we render the default Kustomize overlay and inject the results
into the CSV.

Now that all our CRDs are generated from Go structs, we can leverage
Operator SDK markers to generate OLM descriptors. These descriptors are
required to be accepted into public registries and the Red Hat
Marketplace.

Issue: [ch10384]
Issue: [ch10782]
Issue: [ch11988]