Custom Affinity & Tolerations for Restore Jobs
This change takes custom Affinity and Tolerations fields from the spec
and defines the in a restore job. Users will now be able to ensure that
any/all restore jobs are scheduled according to their specific
requirements and/or constraints (e.g. by using Affinity rules and
Tolerations to influence Pod scheduling).

The dataSource field is supported for both in-place restores and when
restoring from another cluster. Update the dataSource field for your
restore method and the custom settings you define will be added to the
job.

For cloning restores:
```
spec:
  dataSource:
    postgresCluster:
      repoName: repo1
      clusterName: hippo
      clusterNamespace: postgres-operator
      tolerations: {}
      affinity: {}
```

For in-place restores:

```
spec:
  backups:
    pgbackrest:
      restore:
        enabled: true
        repoName: repo1
        tolerations: {}
        affinity: {}
```

Apply suggestions from code review

Co-authored-by: Chris Bandy <bandy.chris@gmail.com>