Consistent name on startup instance
In some environments, caching issues when GETing a cluster from the
client, lead to errors when applying changes to the restore job. Using a
hash of `cluster.Name` and `instanceSet.Name` to generate the startup
instance name ensures that the `StartupInstance` status is set
consistently and the `prepareForRestore` function is idempotent. This
logic is used when using a `dataSource` to restore from another cluster.