OpenShift Auto-detection
Update the operator to automatically detect if it is running in an
OpenShift environment. When the postgres-operator binary starts, it will
check if its being run in an OpenShift environment and pass a value into
the reconciler. When a postgrescluster is reconciled, this value will be
used to set `cluster.Spec.Openshift` if the field is not set by the user
in the spec.