Remove spec.OpenShift references from tutorial doc
Now that auto-detection for OpenShift environments is in place, users
don't have to set the `spec.openshift` field when creating clusters.
Setting `spec.openshift` is still mentioned as a troubleshooting step if
you have pods stuck in the `Pending` phase. All other mentions of this
field have been removed from the tutorial and quickstart docs. If
`spec.openshift` is set in the spec, the value is honored but optional
even when running in OpenShift environments.