Update Error Handling for Volume Functions
The getPVCName(), getRepoPVCNames() and getPGPVCNames() functions now
return any errors encountered when obtaining PersistentVolumeClaim
(PVC) names.  This will ensure errors bubble up to the main reconcile
functions as needed (for instance, to requeue in the event that an
attempt to List PVCs fails due to a transient error), while also
ensuring errors are handled differently that other non-error scenarios
(e.g. currently an error when getting a PVC name produces the same
result as a missing PVC).

Issue: [ch11966]