Allow PostgreSQL to be exposed outside of Kubernetes
This allows one to configure the type of service that routes to the
PostgreSQL primary instance. The type is applied to the HA service to
ensure that the NodePort and/or LoadBalancer always route to exactly one
instance (to avoid split-brain). The primary service is updated to show
the IP address or hostname of the externally provisioned load balancer.

Issue: [ch12074]
Issue: CrunchyData/postgres-operator#2555