Remove "RepoHost" Concept
Removes the concept of a "RepoHost" that is not a "dedicated"
repository host, but rather allows each instance to act as a repo host
that is able to communicate with all other instances (i.e. repo hosts).

With this change, a dedicated repository host Deployment will now
always be created when a "volume" (i.e. non-cloud based) repo is
configured.  Therefore, if only cloud-based repos (S3, GCS or Azure)
are configured, then no dedicated repository host will be created.

Please note that this commit does not yet adjust pgBackRest
configuration, which still might contain configuration for other
instances depending on the current PostgresCluster configuration.