Remove Per-Instance BackRest Config
Removes per-instance pgBackRest configuration file, since pgBackRest
configuration is no longer unique for each instance within the cluster.
Instead, a single "pgbackrest_instance.conf" configuration file is
generated and shared by all instances within the cluster.

This not only simplifies pgBackRest configuration and reconciliation,
but also prevents the need to re-create certain resources in the event
of a failover or switchover (for instance, Jobs that mount pgBackRest
configs for the current primary would would not need to be re-created
when the primary changes due to a switchover or failover).