Custom Database Init Scripts (CRD)
Update the crd to allow users to execute initialization SQL in their
database by pointing to a configmap.

The `DatabaseInitSQL` field has been added to the spec along with two
sub-fields. The first is the name of a ConfigMap containing a SQL file
that will be executed in the database. This ConfigMap must exist in the
same namespace as the cluster. The second field is the ConfigMap data
key that points to the SQL file.

The `DatabaseInitSQL` status has been updated to keep track of the
configmap used to execute SQL in the database. This status is removed
when the DatabaseInitSQL field is removed from the spec.

[ch10815]