Cleanup Bootstrap Secret When rmdata Runs
When a rmdata Job is run to delete a PG cluster, it will now attempt to
remove the pgBackRest bootstrap Secret (if present).  This ensures that
the bootstrap Secret is properly removed after a failed bootstrap
attempt, i.e. when the cluster is deleted and then created again to
reattempt creation of the PG cluster.

Issue: [ch12505]