Add writablePod helper function
This change adds a helper function to the observedInstances type.
Provide a container name to the `writablePod` function and it will
return an Instance that is not terminating, running, and writable along
with the corev1.Pod that is associated with the Instance. Tests are also
added to ensure the function will return nil if no pods meet the
requirements (!terminating, running, writable)