Data Source Now Ignored When Upgrading
A bootstrap Job will now no longer be run if a "pgDataSource" is
present in the pgcluster when it is recreated as part of the'
"pgo upgrade" process.  Specifically, if the "upgrade-in-progress"
annotation is present on the pgcluster, then it is assumed the cluster
is being recreated a part of a PGO upgrade, and any data source logic
is ignored.

Issue: [ch12469]