Pod topology spread constraints
This commit enables the use of "pod topology spread constraints" on
instance, pgBouncer and pgBackRest repo host pods. This feature controls
how pods are spread across your Kubernetes cluster based on user-defined
topology domains. The basic configuration is as follows:

spec:
  topologySpreadConstraints:
  - maxSkew: <integer>
    topologyKey: <string>
    whenUnsatisfiable: <string>
    labelSelector: <object>

where "maxSkew" describes the maximum degree to which Pods can be unevenly
distributed, "topologyKey" is the key that defines a topology in the Nodes'
labels, "whenUnsatisfiable" specifies, when "maxSkew" can't be satisfied,
what action should be taken and "labelSelector" is used to find matching Pods.

Pod topology spread contraints works with related features, such as
PodAffinity/PodAntiAffinity to determine where pods will be placed, but is
not in itself a replacement for these other features.

[ch11731]