Custom Database Init SQL
As part of the reconcile loop, the DatabaseInitSQL field is checked.
When found, the name and fieldPath from this field will be used to get
data from the named Kubernetes ConfigMap. This data will be treated as a
SQL file and executed in the database. The SQL statement(s) are passed
in via stdin and may contain app-psql variables defined in the link
below.

If the psql call returns with exit code 0, the DatabaseInitSQL status
will be set on the cluster. If the call returns with a failure exit
code, the reconcile loop will continue to get the SQL from the ConfigMap
and execute until either the call succeeds or DatabaseInitSQL is removed
from the spec. Unless the app-psql variable `ON_ERROR_STOP` is defined
(using `\set ON_ERROR_STOP` in the SQL file), the psql call will succeed
whenever the last SQL statement succeeds. Failed statements made before
the final statement will be ignored when setting the status.

If the DatabaseInitSQL status is lost or removed and DatabaseInitSQL
still exists in the spec, the SQL will be executed in the database
again. Considering this, SQL statements should be idempotent. When
DatabaseInitSQL is removed from the spec, the DatabaseInitSQL status
will also be removed.

https://www.postgresql.org/docs/current/app-psql.html#APP-PSQL-VARIABLES

[ch10815]