Add initial Kuttl e2e tests
This change adds a quick end-to-end test that use the kuttl test
framework. The layout for kuttl tests is `testing/kuttl/e2e` with
individual tests in subdirectories of the `e2e` directory.

Run test(s) by using the `make check-kuttl` target.

The initial creates a cluster then checks the postgrescluster status and
that the initial backup completes.