Update OpenShift Auto-Detect Logic
Updates the OpenShift auto-detection logic to remove the potential
for false positives.  Specifically, PGO now specifically checks
for the presence of the "SecurityContextConstraints" Kind within the
"security.openshift.io" Group to determine if the Kubernetes cluster
it is running within is an OpenShift cluster.

This prevents false positives by looking for a specific Group/Kind that
will only ever be present in an OpenShift environment (note that the
current logic is looking at a Group suffix that could exist in a
non-OpenShift environment, e.g. if Rook is installed, resulting in a
"replication.storage.openshift.io" Group).

Issue: [sc-12715]