Update Backup Method in Upgrade Docs
The "Backup" method in the "PGO v4 to PGO v5" guide no longer has
specific steps for adjusting the permissions of the PGO v4 pgBackRest
repo for those using NFS or HostPath storage.  This is because the
the specific steps provided do not work consistently in all Kubernetes
envs (e.g. OpenShift will not allow "chown" to be run when using the
"restricted" SCC).  Additionally, the specific permissions applied
could also vary depending on the specific type of storage being
utilized (e.g. NFS or HostPath), as well as how the storage is
configured, provisioned, etc.

Instead, a warning is now displayed for NFS and HostPath users
indicating that they should properly adjust their repo volume
permissions prior to creating a PGO v5 cluster using that repo.
Please note that the specific wording utilized was pulled from the
"Migrate Data Volumes to New Clusters" guide to keep all guidance
consistent.