Reconcile a Basic pgAdmin Deployment
This commit allows for the creation of a pgAdmin deployment. As the initial
implementation of pgAdmin for v5, it reconciles the following:

A Deployment with a single replica only using the latest PGO v4 crunchy-pgadmin image
A Service for accessing pgadmin with a configurable type
A PVC for storing pgAdmin data

Additionally, the Deployment scales to 0 when the cluster is shut down, and the
following items are configurable:

resources, labels, annotation, affinity, priority classes, and
topology spread constraints

Image pull secrets and policies are also supported.

Please note that the current startup scripts within the crunchy-pgadmin container
are being leveraged, but further integration is required before pgAdmin can be used
for managing PostgresClusters.

Finally, when the pgAdmin section is removed from the spec, all pgAdmin resources are
be shut down and removed, including the PVC (retention is based on the PV configuration).

Issue: [sc12517]