Make Patroni Status Non-Pointer
Initially the patroni status only had one purpose meaning it could be
overwritten when its one field (SystemIdentifier) was invalid. Now that
another Patroni status is being added we need to be more careful when
making changes to it. By making this change we can update the status in
both places (i.e. the SystemIdentifier field and the new Switchover
field) without having to worry about instantiating the pointer value. We
can do this because we create the Patroni status on every cluster that
gets created by the operator. There are also some checks throughout the
code that get updated to check for a value in
`Status.Patroni.SystemIdentifier` instead of the existence of
`Status.Patroni`.