CRD Change enabling Ad-Hoc Failovers & Switchovers
The postgrescluter spec now allows users to define a section that will
enable Ad-Hoc Failovers and Switchovers. `Spec.Patroni.Switchover` can
be defined to enable checks for the switchover trigger annotation,
specify a target instance for your switchover or failover, and enable
failovers (when `Spec.Patroni.Switchover.Failover` is false the default
behavior is to perform a Patronictl Switchover).

A new status has been added to keep track of the last switchover
annotation value. When the status matches the current annotation a
switchover will not be performed.