Allow Ad-Hoc Failovers & Switchover
These changes allow users to manually perform a switchover or failover
within their PostgreSQL cluster. Users can specify a target instance
that will become the new primary instance in the cluster. If a target
instance is not provided then Patroni will decided which instance will
be promoted. Note that when running a failover, Patroni requires a
target instance is provided.

PGO will start looking for the `trigger-switchover` annotation when the
Switchover spec (`Spec.Patroni.Switchover.Enabled`) is enabled on your
cluster. If the annotation is found, options defined in the spec will be
used to perform the switchover or failover.

An API call for Patroni failover has been added to facilitate failovers.

If there is an error when trying to change the primary, PGO will be
unable to reconcile the cluster. Errors encountered when reconciling a
patroni switchover/failover will be returned to the main reconcile loop.

Ad-Hoc Switchover & Failover Tutorial