Reconcile pgAdmin4 users through its Python models
pgAdmin4 is a multi-user Flask application that requires one to log in
to use it. We have specifications for PostgreSQL users that we manage.
This commit copies those users and their credentials into pgAdmin4 when
it is the specified user interface for a PostgresCluster.

The default user with Administrator privileges is deactivated.

Issue: [sc-12535]