Change pgAdmin Deployment to StatefulSet
Certain properties of a pgAdmin deployment make it better suited to be
deployed via a StatefulSet rather than a Deployment (e.g. it includes
its own embedded database, and we also want to ensure only a single
replica is running at any given time).  This change therefore updates
the intent generated by PGO for deploying pgAdmin to use a StatefulSet
rather than a Deployment.  The StatefulSet uses the default
"RollingUpdate" update strategy, and is also configured with a proper
"ServiceName" as required for a stable network ID.

[sc-12915]