PodDisruptionBudgets for Instance Sets & PGBouncer
Reconcile Pod Disruption Budget's (PDB) for PG Instance sets and
PGBouncer deployments. This ensures a minimum number of PG Instance Pods
in a set and replicas for PGBouncer will remain running during voluntary
disruptions of the Kubernetes cluster.

PDBs are configured by setting the `minAvailable` field in the spec
(`spec.instances.minAvailable` for instances and
`spec.proxy.pgbouncer.minAvailable` for PGBouncer). If the field is set
to 0 a PDB will not be created and any existing PDBs for that resource
will be deleted.

If `minAvailable` is not defined a default value will be set. If your
resource has more than one replica a PDB with a minAvailable of 1 will
be created.