Allow for seamless upgrade to new AWS S3 CA bundle
This updates the autodetection logic to add the new AWS S3 CA
bundle to the general PGO Secret, which is then applied to clusters
on upgrade.

The logic is such that it will only overwrite the default template
if it is unmodified, i.e. it is using the CA bundle that is
provided.