Manage pgAdmin server connections
This commit creates server connections based on the users defined
for the PostgresCluster. With this update, each user is able to
log in with their unique password credential and manage their
Postgres database sessions. Relevant changes, such as port or
password, are handled automatically, but in some cases may require
reloading the page.

Issue: [sc-12538]