Allow Scaledown of Original Primary After Upgrade
Updates the rmdata process to simply ignore any isNotFound errors when
attempting to delete a pgreplica for a scaledown request.
Specifically, the pgreplica deletion logic is no longer concerned with
the replica name matching the cluster scope when determining whether or
not to ignore an isNotFound error, since the there are scenarios where
we might expect this error even though the replica name does not match
the scope (for instance following a upgrade, where the original
primary might have a name that does not match the scope, e.g. if a
failover has previously occurred prior to the upgrade).

And since the goal of running the rmdata process is to delete any/all
replica resources, including the pgreplica itself, if the replica is
already gone for some reason (e.g. if someone manually deletes it
before rmdata has a chance to, or if the original primary is being
scaled down, which will simply not have an associated pgreplica), the
rmdata process can continue with attempting to cleanup all other
replica resources to ensure the replica is fully removed.

[sc-13225]