Limit Scope of OpenShift Auto Detect Discovery
The OpenShift auto-detection logic has been updated to only check for
server resources (and therefore the "SecurityContextConstraints" kind)
in the "security.openshift.io" group.  This is done to prevent a single
broken (and related) API in the Kubernetes cluster from breaking the
OpenShift detection logic, as can occur when checking all server groups
and resources.

Issue: [sc-13205]