Add an internal test helper that compares to YAML
Multiple packages have copies of the same function, "marshalMatches".
A few packages used an equivalent combination of functions,
"marshalEquals" with "strings.Trim".

Move the implementation to a common package, and replace the lesser-used
"marshalEquals" with it. Replacing every call to "marshalMatches" is
very likely to conflict with other work, so leave those names in place
for now.