Stop writing ExternalIPs for <cluster>-primary service
For convenience, if a cluster was created with `spec.service.type` `LoadBalancer`,
we updated the <cluster>-primary Service with the ExternalIPs and ExternalName.
However, this has led to a problem in GKE auto-pilot clusters, where
ExternalIPs are not allowed due to security concerns.
(See https://cloud.google.com/anthos/clusters/docs/security-bulletins#gcp-2020-015).
This PR removes the code that added the ExternalIPs and ExternalName to the
<cluster>-primary Service.

Issue [sc-13257]