Update pgAdmin SecurityContext
Updates the security context for the pgAdmin StatefulSet to include
the proper fsGroup and supplementalGroups settings.  Specifically,
fsGroup is now set to 26 if not running in OpenShift, while any
supplemental groups defined in the spec are also applied.  This ensures
mounted  pgAdmin volumes are writeable as required by the pgAdmin
application.

Issue: [sc-13287]